backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["hamail <mhamail1223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
pytest = "^8.3.2"
httpx = "^0.27.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
sqlalchemy = "^2.0.32"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
alembic = "^1.13.2"

python-dotenv = "^1.0.1"

psycopg = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

